Define detectMouseWheelZoom function
